pyproject.toml
[tool.poetry]
name = "gaming-website-monorepo"
version = "0.1.0"
description = "Monorepo for gaming library microservices project"
authors = ["Your Name <your@email.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
sqlalchemy = "^2.0.41"
alembic = "^1.16.4"
dotenv = "^0.9.9"
psycopg = {extras = ["binary"], version = "^3.2.9"}
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
black = "*"
pylint = "*"
pre-commit = "*"
isort = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
